Improve loglik calculation with symbolic deformation

+++ src/exclusivity.hpp
@@ -19,7 +19,7 @@ class Exclusivity {
   public:
     Exclusivity(std::istream& infile, const size_t g, const size_t n=65535);
     Exclusivity(const std::vector<std::string>& names,
-          const Eigen::MatrixXd& genotypes,
+          const Eigen::MatrixXi& genotypes,
           const size_t grid_density,
           const size_t max_results=65535):
           names_(names),
@@ -30,7 +30,7 @@ class Exclusivity {
     void run(const std::string& outfile="/dev/stdout");
 
     const std::vector<std::string>& names() const {return names_;}
-    const Eigen::MatrixXd& genotypes() const {return genotypes_;}
+    const Eigen::MatrixXi& genotypes() const {return genotypes_;}
     std::ostream& write_genotypes(std::ostream&, const bool header=true) const;
     std::ostream& write_results(std::ostream&, const bool header=true) const;
 
@@ -39,7 +39,7 @@ class Exclusivity {
     /////1/////////2/////////3/////////4/////////5/////////6/////////7/////////
   private:
     const std::vector<std::string> names_;
-    const Eigen::MatrixXd genotypes_;
+    const Eigen::MatrixXi genotypes_;
     const size_t grid_density_;
     const size_t max_results_;
     std::multimap<double, std::vector<double>> results_;
