+++ src/program.cpp
@@ -106,7 +106,7 @@ Program::Program(const std::vector<std::string>& arguments) {HERE;
 void Program::run() {HERE;
     wtl::Fin fin(INFILE);
     const auto names = wtl::read_header(fin);
-    const auto genotypes = wtl::eigen::read_matrix<double>(fin, names.size());
+    const auto genotypes = wtl::eigen::read_matrix<int>(fin, names.size());
     fin.close();
     // Model model(names, genotypes, GRID_DENSITY, MAX_RESULTS);
     // model.run(THRESHOLD, INTERCEPT, OUTFILE);
